feat: 合并 interface 目录
