+++ ui/src/locales/zh-CN.ts
@@ -25,9 +25,8 @@ export default {
     'project.members': '项目成员',
     'project.invite': '邀请成员',
 
-    'interface': '接口',
+    'interface': '接口管理',
     'interface1': '接口旧',
-    'interfaceV2': '接口管理', // 新版 接口模块
     'projectSetting': '项目设置', // 新版 接口模块
     'scenario': '场景',
     'scenario.edit': '场景编辑',
