+++ ui/src/views/interface/components/Drawer/InterfaceBasicInfo.vue
