--- ui/src/views/interfaceV2/data.d.ts
@@ -1,56 +1,0 @@
-interface InterfaceListReqParams {
-    "prjectId"?: number,
-    "page"?: number,
-    "pageSize"?: number,
-    "status"?: number,
-    "userId"?: number,
-    "title"?: string
-}
-
-interface SaveInterfaceReqParams {
-    project_id?: number,
-    serveId?: number,
-    title?: string,
-    path?: string
-}
-
-
-export interface Interface {
-    id: number;
-    name: string;
-    desc: string;
-}
-
-export interface QueryResult {
-    list: Interface[];
-    pagination: PaginationConfig;
-}
-
-export interface QueryParams {
-    keywords: string,
-    enabled: string,
-    page: number,
-    pageSize: number,
-}
-
-export interface PaginationConfig {
-    total: number;
-    current: number;
-    pageSize: number;
-    showSizeChanger: boolean;
-    showQuickJumper: boolean;
-}
-
-
-export interface filterFormState {
-    status: string | null,
-    createUser: string | null,
-    title: string | null,
-}
-
-
-interface NewInterfaceFormState {
-    title: string;
-    parentId: string | null;
-    description: string | undefined;
-}
