+++ ui/src/views/projectSetting/components/ServiceSetting/Component.vue
@@ -144,7 +144,6 @@ import {
   watch
 } from 'vue';
 import {CheckOutlined, EditOutlined} from '@ant-design/icons-vue';
-// import {getYaml} from "@/views/interfaceV2/service";
 import {
   saveSchema,
   deleteSchema,
